Adapt CLI to new service model
