+++ pkg/service/controlservice.proto
@@ -17,7 +17,16 @@ message QuitResponse {
   bool success = 1;
 }
 
+message SwitchPhaseMessage {
+  int32 phase = 1;
+}
+
+message SwitchPhaseResponse {
+  bool success = 1;
+}
+
 service ControlService {
   rpc Log(LogMessage) returns (LogResponse) {}
   rpc Quit(QuitMessage) returns (QuitResponse) {}
+  rpc SwitchPhase(SwitchPhaseMessage) returns (SwitchPhaseResponse) {}
 }
